Post Finish with Semester Pre Improvements
Finished with project! However, I still want to work on: 
•	Speed
o	The computer has to do many calculations (For every pixel, it has to do a calculation for at least the number of objects in the world). When you start increasing the number of objects in the 3d world, the computer renders these objects must slower. Traditionally, this problem is resolved by sending the calculations that need to be ran to the graphics card and allowing the graphics card to do these types of calculations. So instead of the processor doing these calculations one after another, the graphics card can likely do all the calculations at the same time and it should be exponentially faster. 
•	Direction 
o	Program it so that the user can look in whatever direction he/she wants to. Currently it is set up so the user can only look at things that are only in the –x direction. This type of math would involve the user of “quaternions" which are like a bigger badder version of complex n